change productPage linking
